crypto-advisor/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
gunicorn==21.2.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
